Avoid undefined behavior with left-shift of negative ints.
Caught by UBSan when testing the library using the provided gme_player
sample app.

While this has not caused improper code to be generated, it's still better to
stay ahead of the compiler writers...

Interestingly, part of this was done back in 2013 (commit
6e1544d79a559cee6ff6b98efcc3df41ff27df66) in response to a PVS Studio
check I'd happened to run across, see
http://www.viva64.com/en/examples/V610/.

But I'd missed this one back then when grepping for it, probably due to
the different variable name. :(

+++ gme/Spc_Dsp.cpp
@@ -498,8 +498,9 @@ void Spc_Dsp::run( int clock_count )
 					// Decode four samples
 					for ( end = pos + 4; pos < end; pos++, nybbles <<= 4 )
 					{
-						// Extract upper nybble and scale appropriately
-						int s = ((int16_t) nybbles >> right_shift) << left_shift;
+						// Extract upper nybble and scale appropriately. Every cast is
+						// necessary to maintain correctness and avoid undef behavior
+						int s = int16_t(uint16_t((int16_t) nybbles >> right_shift) << left_shift);
 						
 						// Apply IIR filter (8 is the most commonly used)
 						int const filter = brr_header & 0x0C;
